Improve configuration properties meta-data
This commit changes the type of the `base-dir` property to `Path` as
the standard `ConversionService` is able to handle it from a `String`
value. This has the advantage of rejecting any invalid String value
at bootstrap time rather than failing later.

Also, this commit documents the property so that the description shows
up in IDEs that support code assistance for configuration properties.